requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.2
langserve>=0.2.2
faiss-cpu>=1.8.0
pandas>=2.2.0
python-dotenv>=1.0.0
redis>=5.0.0
requests>=2.31.0
qdrant-client>=1.9.0
langchain-qdrant>=0.1.2
streamlit>=1.36.0

